Removed display_module.h
