+++ app/src/ui/screens/scr_bpt.c
@@ -8,7 +8,6 @@
 #include <app_version.h>
 #include <zephyr/logging/log.h>
 
-#include "display_module.h"
 #include "sampling_module.h"
 #include "hw_module.h"
