+++ app/src/ui/screens/scr_clock_small.c
@@ -9,7 +9,6 @@
 #include <zephyr/logging/log.h>
 #include <zephyr/drivers/rtc.h>
 
-#include "display_module.h"
 #include "sampling_module.h"
 
 #include "ui/move_ui.h"
