+++ app/src/ui/screens/scr_home.c
@@ -6,7 +6,6 @@
 #include <zephyr/logging/log.h>
 #include <zephyr/drivers/rtc.h>
 
-#include "display_module.h"
 #include "ui/move_ui.h"
 
 lv_obj_t *scr_home;
