+++ app/src/ui/screens/scr_hrv_scatter.c
@@ -9,7 +9,6 @@
 #include <app_version.h>
 #include <zephyr/logging/log.h>
 
-#include "display_module.h"
 #include "sampling_module.h"
 
 #include "ui/move_ui.h"
