+++ app/src/ui/screens/scr_vitals.c
@@ -10,10 +10,7 @@
 #include <app_version.h>
 #include <zephyr/logging/log.h>
 
-
-#include "display_module.h"
 #include "sampling_module.h"
-
 #include "ui/move_ui.h"
 
 // GUI Labels
